==> prev_reported_safety_associations/src/2_Bioportal_annotation.ipynb <==
import urllib.request, urllib.error, urllib.parse
import json
import os
import pandas as pd
import time

pd.set_option('display.max_colwidth', 300)

basedir = '/scratch/ias41/ae_code/prev_reported_safety_associations'
with open('/scratch/ias41/ucc-fs-nethome/bioportal_api_key.txt', 'r') as f:
    API_KEY = f.read().strip('\n')

def get_json(url):
    opener = urllib.request.build_opener()
    opener.addheaders = [('Authorization', 'apikey token=' + API_KEY)]
    return json.loads(opener.open(url).read())

REST_URL = "http://data.bioontology.org"
def get_bioportal_annotations(text):
    """Return codes of MedDRA terms annotated by BioPortal Annotate Service.
    kwargs: text -- test to annotate"""
    
    annotations = get_json(REST_URL + "/annotator?text=" + urllib.parse.quote(text) + '&ontologies=MEDDRA&longest_only=true')
    
    identifiers = set()
    for result in annotations:
        identifier = result['annotatedClass']['@id'].split('/')[-1]
        identifiers.add(identifier)
    
    return(identifiers)

# Open previously formatted adverse effects extracted from publications
formatted_effects = pd.read_excel(basedir + '/data/safety_original2formatted_effects.xls')

formatted_effects.head()

# find all unique AEs occurring
all_aes = set(list(formatted_effects['Formatted effect']))

%%time
# Find annotations from BioPortal
ae_tuples = list()

for effect in all_aes:
    annotations = get_bioportal_annotations(effect)
    time.sleep(0.02)
    if len(annotations) < 1:
        ae_tuples.append((effect, 'Nothing found'))
        continue
    for annotation in annotations:
        ae_tuples.append((effect, annotation))

ae_df = pd.DataFrame.from_records(ae_tuples)
ae_df.columns = ['Original formatted effect', 'MedDRA ID']
len(ae_df)

# Prepare file for submitting to MedDRA browser hierarchy analysis
# Save to file for hierarchy analysis in MedDRA to find LLT > PT
ae_df_mapped = ae_df.loc[(ae_df['MedDRA ID']!='Nothing found')&(~ae_df['MedDRA ID'].str.contains('T')),:]
ae_df_mapped.columns=['Row ID', 'Code']
ae_df_mapped['Term'] = ''
ae_df_mapped[['Row ID', 'Term', 'Code']].to_excel(basedir + '/data/formatted_effects_BioPortal_annotated.xls', index=False)

# Number of terms with at least one annotation
len(ae_df_mapped['Row ID'].drop_duplicates())

# Do MedDRA hierarchy analysis in MedDRA Web-based browser (WBB) and manual inspection of the results
