==> prev_reported_safety_associations/src/3_annotate_rest.ipynb <==
import pandas as pd

pd.set_option('display.max_colwidth', 300)

basedir = '/scratch/ias41/ae_code/prev_reported_safety_associations'

all_effects = pd.read_excel(basedir + '/data/safety_original2formatted_effects.xls', sep='\t')

all_effects.head()

# Total number of unique terms
len(all_effects['Formatted effect'].drop_duplicates())

WBB_effects = pd.read_excel(basedir + '/data/formatted_effects_WBB_hierarchy_output_inspected.xlsx')
WBB_effects_accepted = WBB_effects.loc[WBB_effects['Manual inspection accepted']=='y',:]
WBB_effects_accepted.drop_duplicates(inplace=True)
WBB_effects_accepted['PT Code'] = WBB_effects_accepted['PT Code'].astype(int)

# Number of currently annotated effects
len(WBB_effects_accepted['Row ID'].drop_duplicates())

# Make file with terms not yet annotated
not_annotated_terms = pd.DataFrame(all_effects.loc[~all_effects['Formatted effect'].isin(WBB_effects_accepted['Row ID'].tolist()),'Formatted effect'].drop_duplicates())

# Number of unique terms not_annotated
len(not_annotated_terms)

# Save two files, one with un-annotated terms, other convenience file with annotated terms to use as reference
not_annotated_terms.sort_values(by='Formatted effect').to_csv(basedir + '/data/formatted_effects_WBB_not_annotated.txt', sep='\t', index=False)
WBB_effects_accepted[['Row ID', 'PT', 'PT Code']].drop_duplicates().sort_values(by='Row ID').to_csv(basedir + '/data/formatted_effects_WBB_accepted.txt', sep='\t', index=False)
