==> prev_reported_safety_associations/src/4_compile_PT_effects.ipynb <==
"""Combine the automated and manual annotations"""

import pandas as pd

basedir = '/scratch/ias41/ae_code/prev_reported_safety_associations'

all_formatted_effects = pd.read_excel(basedir + '/data/safety_original2formatted_effects.xls')

WBB_annotations = pd.read_csv(basedir + '/data/formatted_effects_WBB_accepted.txt', sep='\t')
manual_annotations_original = pd.read_csv(basedir + '/data/formatted_effects_manually_annotated.txt', sep='\t')
manual_annotations = manual_annotations_original.loc[~manual_annotations_original['MedDRA PT code'].isnull(),:].copy()
manual_annotations['MedDRA PT code'] = manual_annotations['MedDRA PT code'].astype(int)
manual_annotations.drop(labels=['Other level term', 'Parsing mistake', 'Comments', 'Other level code', 'Other level'], inplace=True, axis=1)
manual_annotations.columns = ['Formatted effect', 'PT Code', 'PT']

WBB_merge = all_formatted_effects.merge(WBB_annotations, left_on='Formatted effect', right_on='Row ID')
WBB_merge.drop(labels='Row ID', axis=1, inplace=True)

manual_merge = all_formatted_effects.merge(manual_annotations, on='Formatted effect')

total_merged = pd.concat([WBB_merge, manual_merge], sort=False)

total_merged.columns = ['Accession', 'ChEMBL target name', 'Original effect',
       'Formatted effect', 'PubMed ID', 'Reference', 'Annotated MedDRA PT',
       'Annotated MedDRA PT Code']
total_merged.sort_values(by=['ChEMBL target name', 'Formatted effect'], inplace=True)

total_merged.to_excel(basedir + '/data/safety_meddra_annotated_effects.xls', index=False)

### Save manually annotated HLTs

hlt_manual = manual_annotations_original.loc[manual_annotations_original['Other level']=='HLT',:]
hlt_manual.drop(labels=['Parsing mistake', 'Comments', 'MedDRA PT code', 'MedDRA PT','Other level'], inplace=True, axis=1)
hlt_manual_merged = all_formatted_effects.merge(hlt_manual, on='Formatted effect')
hlt_manual_merged['Other level code'] = hlt_manual_merged['Other level code'].astype(int)

hlt_manual_merged.columns = ['Accession', 'ChEMBL target name', 'Original effect',
       'Formatted effect', 'PubMed ID', 'Reference', 'Annotated MedDRA HLT Code',
       'Annotated MedDRA HLT']

hlt_manual_merged.to_excel(basedir + '/data/safety_meddra_manually_annotated_hlt_effects.xls', index=False)

### Save a copy of the PTs for hierarchy analysis in MedDRA browser

hierarchy_selection = total_merged[['Annotated MedDRA PT', 'Annotated MedDRA PT Code']].drop_duplicates().reset_index(drop=True)

hierarchy_selection['Row ID'] = ''
hierarchy_selection.columns = ['Term', 'Code', 'Row ID']

hierarchy_selection[['Row ID', 'Term', 'Code']].to_excel(basedir + '/data/safety_meddra_annotated_effects_for_hierarchy_input.xls', index=False)
